add device

+++ vllm/worker/tpu_worker.py
@@ -115,8 +115,10 @@ class TPUWorker(LoraNotSupportedWorkerBase, LocalOrDistributedWorkerBase):
         head_size = self.model_config.get_head_size()
         num_kv_heads = self.model_config.get_num_kv_heads(self.parallel_config)
 
-        kv_caches = [(torch.tensor([], dtype=torch.int, device=self.device),
-                      torch.tensor([], dtype=torch.int, device=self.device))
+        kv_caches = [(torch.tensor([], dtype=torch.float32,
+                                   device=self.device),
+                      torch.tensor([], dtype=torch.float32,
+                                   device=self.device))
                      for _ in range(num_layers)]
         self.model_runner._dummy_run(
             batch_size=1,
